requirements.txt
pyinstaller
requests
pycryptodome
pyjwt